% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_OCR_subtype.R
\name{calculate_OCR_subtype}
\alias{calculate_OCR_subtype}
\title{Calculate OCR subtype}
\usage{
calculate_OCR_subtype(
  data_min = data_min,
  bin_list = bin_list,
  return_type = "df"
)
}
\arguments{
\item{data_min}{a data frame}

\item{bin_list}{a list}

\item{return_type}{df" or "plot"}
}
\value{
mito_Base_Respiration, mito_Proton_Leak, mito_ATP_Production, mito_Max_Respiration, mito_Non_Mito_Oxygen_Consumption, mito_Spare_Respiratory_Capacity
}
\description{
Calculate OCR subtype
}
\examples{
calculate_OCR_subtype <- function(data_min=data_min, bin_list=bin_list, return_type="df")
calculate_OCR_subtype <- function(data_min=data_min, bin_list=bin_list, return_type="plot")

}
