% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{calculate_well_id}
\alias{calculate_well_id}
\title{Calculate well_id based on row and col}
\usage{
calculate_well_id(Row, Col)
}
\arguments{
\item{Row}{character A, B, C, D}

\item{Col}{character 1,2,3,4,5,6}
}
\value{

}
\description{
Calculate well_id based on row and col
}
\examples{

}
