% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_func.R
\name{convert_layout_to_long}
\alias{convert_layout_to_long}
\title{Convert the layout of a plate from wide format to long format}
\usage{
convert_layout_to_long(csvfile)
}
\arguments{
\item{csvfile}{input by user}
}
\value{

}
\description{
Convert the layout of a plate from wide format to long format
}
\examples{
file = system.file("extdata", "layout.csv", package="rSeahorse")
convert_layout_to_long(csvfile = file)
}
