% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group.R
\name{create_group}
\alias{create_group}
\title{Title}
\usage{
create_group()
}
\value{
a data frame with columns: Well, Group, Col, Row, Treatment
}
\description{
Title
}
\examples{
group_info <-  create_group()
}
