% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_seahorse.R
\name{import_seahorse}
\alias{import_seahorse}
\title{Import seahorse data from am excel file , clean and organize data for viewing and plotting}
\usage{
import_seahorse(file = file, group_info = group_info)
}
\arguments{
\item{file}{an excel file with worksheet: "Rate Data", "Time Events", which is exported by the seahorse software}

\item{group_info}{group names and  for each group, how many wells are included}
}
\value{
a list  {data, datta_min, event lable, well_label, sheet_names}
}
\description{
Import seahorse data from am excel file , clean and organize data for viewing and plotting
}
\examples{




}
