% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_OCR_subtype.R
\name{make_time_bin}
\alias{make_time_bin}
\title{Tittle}
\usage{
make_time_bin(event_label)
}
\arguments{
\item{event_label}{a data frame including columns: Time, Event, Prarmeter, minutes, treatment, y}
}
\value{
a list
}
\description{
Tittle
}
\examples{

}
