% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seahorse.R
\name{plot_seahorse}
\alias{plot_seahorse}
\title{Title}
\usage{
plot_seahorse(
  data_min = data_min,
  wells_exclude = wells_exclude,
  event_label = event_label,
  group_info = group_info,
  size_group
)
}
\arguments{
\item{data_min}{data with minutes}

\item{wells_exclude}{a vector}

\item{event_label}{label event}

\item{group_info}{a named vector, name is the labels, value is the levels}

\item{size_group}{a number, for calculation SEM}
}
\value{

}
\description{
Title
}
\examples{
plot_seahorse(data_min, well_exclude, event_label, group_info, size_group =6 )

}
