% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view.R
\name{qview_seahorse}
\alias{qview_seahorse}
\title{An overview of the whole plate}
\usage{
qview_seahorse(data_min, well_label)
}
\arguments{
\item{list}{returned by import_seahourse}
}
\value{

}
\description{
An overview of the whole plate
}
\examples{
qview_seahorse(data_min, well_label)

}
